--- 1/posts/index.js
@@ -1,4 +1,0 @@
-import BlogPagination, { getStaticProps } from "./page/[slug]";
-
-export { getStaticProps };
-export default BlogPagination;
