+++ app/posts/page.js
@@ -1,0 +1,4 @@
+import BlogPagination, { generateStaticParams } from "./page/[slug]/page";
+
+export { generateStaticParams };
+export default BlogPagination;
